[android] Move FontManager interface to unimodules-core
