+++ packages/@unimodules/core/android/src/main/java/org/unimodules/core/interfaces/services/FontManager.java
@@ -1,0 +1,7 @@
+package org.unimodules.core.interfaces.services;
+
+import android.graphics.Typeface;
+
+public interface FontManager {
+  void setTypeface(String fontFamilyName, int style, Typeface typeface);
+}
